# perf(@angular-devkit/build-angular): add vendor chunking to server builder