--- a/goldens/public-api/angular_devkit/build_angular/index.md
+++ b/goldens/public-api/angular_devkit/build_angular/index.md
@@ -269,6 +269,7 @@ export interface ServerBuilderOptions {
     statsJson?: boolean;
     stylePreprocessorOptions?: StylePreprocessorOptions_3;
     tsConfig: string;
+    vendorChunk?: boolean;
     verbose?: boolean;
     watch?: boolean;
 }