--- a/packages/schematics/angular/migrations/update-15/update-workspace-config.ts
+++ b/packages/schematics/angular/migrations/update-15/update-workspace-config.ts
@@ -18,8 +18,12 @@ export default function (): Rule {
           continue;
         }
 
-        for (const [, options] of allTargetOptions(target)) {
+        for (const [name, options] of allTargetOptions(target)) {
           delete options.bundleDependencies;
+
+          if (name === 'development') {
+            options.vendorChunk ??= true;
+          }
         }
       }
     }